LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/scrollend-event-not-fired-on-no-scroll-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_true: no scroll, so no scrollend expected expected true got false
target
doc

Harness Error (FAIL), message = Unhandled rejection: assert_true: no scroll, so no scrollend expected expected true got false

PASS No scroll via wheel on div shouldn't fire scrollend.
PASS No scroll via keys on div shouldn't fire scrollend.
FAIL No scroll via wheel on document shouldn't fire scrollend. assert_true: no scroll, so no scrollend expected expected true got false
PASS No scroll via keys on document shouldn't fire scrollend.
